backend/requirements.txt
pip~=20.1
pytz~=2020.1
marvelous~=1.0.4
Django~=3.0.6
sqlparse~=0.3.1
asgiref~=3.2.7
certifi~=2020.4.5.1
chardet~=3.0.4
Markdown~=3.2.2
idna~=2.9
urllib3~=1.25.9
marshmallow~=3.6.0
requests~=2.23.0
setuptools~=46.1.3
django-filter~=2.2.0
djangorestframework~=3.11.0
python-dotenv~=0.13.0